+++ src/app/(auth)/app/(sidebar)/customers/_components/modals.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-hooks/exhaustive-deps */
 'use client';
 
 import { useEffect, useRef, useState } from 'react';
@@ -220,7 +221,7 @@ export function CustomerFormModal({ isOpen, onClose, title, submitText = 'Xác n
         staffId: initialData?.staffId || (!hasFullViewRole && user ? user.id : ''),
         type: initialData?.type || '',
       });
-      // eslint-disable-next-line react-hooks/set-state-in-effect
+       
       setSelectedImages([]);
 
       const getFullImageUrl = (path: string) => {